Fixed bug of outputing euros without decimals and having overflow after absurd amount of cents in case of int

+++ Currency.h
@@ -62,7 +62,7 @@ public:
 	// Operator Overloads END #####
 
 private:
-	int Amount{ 0 };
+	long long Amount{ 0 };
 };
